refactor(Picker): replacing Picker with PickerV2 components
