--- packages/react-components/src/components/Picker/Picker.mdx
@@ -1,38 +1,0 @@
-import { Meta, Title, Canvas, ArgTypes } from '@storybook/blocks';
-import { Figma } from '@storybook/addon-designs/blocks';
-import * as Picker from './Picker.stories';
-
-<Meta of={Picker} />
-
-<Title>Picker</Title>
-
-[Intro](#Intro) | [Component API](#ComponentAPI) | [Content Spec](#ContentSpec)
-
-## Intro <a id="Intro" />
-
-<Canvas of={Picker.Default} sourceState="none" />
-
-#### Example implementation
-
-```jsx
-<Picker
-  options={[
-    { key: 'one', name: 'Option one' },
-    { key: 'two', name: 'Option two' },
-  ]}
-  selected={selected}
-  onSelect={(items) => setItems(items)}
-/>
-```
-
-## Component API <a id="ComponentAPI" />
-
-<ArgTypes of={Picker.Default} />
-
-## Content Spec <a id="ContentSpec" />
-
-<Figma
-  showLink
-  height="100%"
-  url="https://www.figma.com/file/9FDwjR8lYvincseDkKypC4/Component-Documentations?type=design&node-id=731-53240&mode=dev"
-/>
