+++ packages/react-components/src/components/Picker/Picker.tsx
@@ -1,98 +1,38 @@
 import * as React from 'react';
 
-import cx from 'clsx';
+import { cx } from '@emotion/css';
+import {
+  autoUpdate,
+  flip,
+  shift,
+  FloatingPortal,
+  offset,
+  size as floatingSize,
+  useClick,
+  useDismiss,
+  useFloating,
+  useInteractions,
+  useListNavigation,
+  useRole,
+  FloatingNode,
+  useFloatingNodeId,
+} from '@floating-ui/react';
+import * as ReactDOM from 'react-dom';
 
-import { Size } from 'utils';
-
-import { KeyCodes } from '../../utils/keyCodes';
-
-import { SELECT_ALL_OPTION_KEY } from './constants';
-import { PickerList } from './PickerList';
-import { Trigger } from './Trigger';
-import { TriggerBody } from './TriggerBody';
-import { IPickerListItem, PickerType } from './types';
+import { PickerList } from './components/PickerList';
+import { PickerTrigger } from './components/PickerTrigger';
+import { PickerTriggerBody } from './components/PickerTriggerBody';
+import { findIndicesWhere } from './helpers';
+import { IPickerListItem, IPickerProps } from './types';
 
 import styles from './Picker.module.scss';
 
-const baseClass = 'picker';
-
-export interface IPickerProps {
-  /**
-   * Specify the custom id
-   */
-  id?: string;
-  /**
-   * The CSS class for picker container
-   */
-  className?: string;
-  /**
-   * Specify whether the picker should be disabled
-   */
-  disabled?: boolean;
-  /**
-   * Specify whether the picker should be in error state
-   */
-  error?: boolean;
-  /**
-   * Array of picker options
-   */
-  options: IPickerListItem[];
-  /**
-   * Array of picker selected options
-   */
-  selected?: IPickerListItem[] | null;
-  /**
-   * Specify the picker size
-   */
-  size?: Size;
-  /**
-   * Specify the placeholder for search input
-   */
-  placeholder?: string;
-  /**
-   * Specify whether the option select is required
-   */
-  isRequired?: boolean;
-  /**
-   * Text if no search result were found
-   */
-  noSearchResultText?: string;
-  /**
-   * Text for `select all` option which will be visible if defined in multi select mode
-   */
-  selectAllOptionText?: string;
-  /**
-   * Set `multi` to specify whether the picker should allow to multi selection
-   */
-  type?: PickerType;
-  /**
-   * Set to disable search input
-   */
-  searchDisabled?: boolean;
-  /**
-   * Set to hide clear selection button
-   */
-  hideClearButton?: boolean;
-  /**
-   * Will open picker on component initialization
-   */
-  openedOnInit?: boolean;
-  /**
-   * Test id passed to the picker trigger element
-   */
-  ['data-testid']?: string;
-  /**
-   * Callback called after item selection
-   */
-  onSelect: (selectedItems: IPickerListItem[] | null) => void;
-  /**
-   * Clears the search input after item select
-   */
-  clearSearchAfterSelection?: boolean;
-}
+const overflowPadding = 10;
 
 export const Picker: React.FC<IPickerProps> = ({
+  id,
   className,
+  listClassName,
   disabled,
   error,
   options,
@@ -108,114 +48,25 @@ export const Picker: React.FC<IPickerProps> = ({
   openedOnInit = false,
   clearSearchAfterSelection,
   onSelect,
+  floatingStrategy,
+  useDismissHookProps,
+  useClickHookProps,
   ...props
 }) => {
-  const [isListOpen, setIsListOpen] = React.useState<boolean>(openedOnInit);
+  const [open, setOpen] = React.useState(openedOnInit);
+  const [pointer, setPointer] = React.useState(false);
+  const [selectedKeys, setSelectedKeys] = React.useState<string[]>(
+    () => selected?.map(({ key }) => key) || []
+  );
+  const [activeIndex, setActiveIndex] = React.useState<number | null>(null);
   const [searchPhrase, setSearchPhrase] = React.useState<string | null>(null);
-  const triggerRef = React.useRef<HTMLDivElement>(null);
+  const [maxHeight, setMaxHeight] = React.useState(400);
+  const listElementsRef = React.useRef<Array<HTMLElement | null>>([]);
+  const nodeId = useFloatingNodeId();
 
-  const mergedClassNames = cx(styles[baseClass], className);
-
-  React.useEffect(() => {
-    if (isListOpen) {
-      const onDocumentClick = (e: MouseEvent) => {
-        if (!triggerRef.current?.contains(e.target as Element)) {
-          setIsListOpen(false);
-        }
-      };
-
-      const onKeyDown = (e: KeyboardEvent) => {
-        if (e.key === KeyCodes.tab) {
-          setIsListOpen(false);
-        }
-      };
-
-      document.addEventListener('mousedown', onDocumentClick);
-      document.addEventListener('keydown', onKeyDown);
-
-      return () => {
-        document.removeEventListener('mousedown', onDocumentClick);
-        document.addEventListener('keydown', onKeyDown);
-      };
-    } else {
-      setSearchPhrase(null);
-    }
-  }, [isListOpen]);
-
-  const handleTrigger = (e: React.MouseEvent | KeyboardEvent) => {
-    const target = e.target as Element;
-
-    if (disabled || target.getAttribute('data-dismiss-icon')) {
-      return;
-    }
-
-    setIsListOpen((prev) => !prev);
-  };
-
-  const handleOnClose = () => {
-    setIsListOpen(false);
-  };
-
-  const handleSelect = (item: IPickerListItem) => {
-    if (type === 'single') {
-      setIsListOpen(false);
-
-      return onSelect([item]);
-    }
-
-    const selectedItemKey = item.key;
-    const currentSelectedItemsKeys = selectedItemsKeys;
-
-    if (!currentSelectedItemsKeys) {
-      return onSelect([item]);
-    }
-
-    const newSelectedItemsKeys = currentSelectedItemsKeys.includes(
-      selectedItemKey
-    )
-      ? currentSelectedItemsKeys.filter((key) => key !== selectedItemKey)
-      : currentSelectedItemsKeys.concat(selectedItemKey);
-
-    if (newSelectedItemsKeys?.length === 0) {
-      return onSelect(null);
-    }
-
-    const newSelectedItems = options.filter((item) =>
-      newSelectedItemsKeys.includes(item.key)
-    );
-
-    onSelect(newSelectedItems);
-  };
-
-  const isItemSelectable = (item: IPickerListItem) =>
-    !item.disabled && !item.groupHeader && item.key !== SELECT_ALL_OPTION_KEY;
-
-  const handleSelectAll = () => {
-    setIsListOpen(false);
-
-    const itemsToSelect = items.filter(isItemSelectable);
-
-    onSelect(itemsToSelect);
-  };
-
-  const handleClear = () => {
-    setIsListOpen(false);
-    onSelect(null);
-  };
-
-  const handleOnFilter = (text: string) => setSearchPhrase(text);
-
-  const handleItemRemove = (item: IPickerListItem) => {
-    const newSelectedItems = selected
-      ? selected.filter((selectedItem) => selectedItem !== item)
-      : null;
-
-    if (newSelectedItems?.length === 0) {
-      return onSelect(null);
-    }
-
-    onSelect(newSelectedItems);
-  };
+  if (!open && pointer) {
+    setPointer(false);
+  }
 
   const items = React.useMemo<IPickerListItem[]>(() => {
     if (!searchPhrase) {
@@ -233,57 +84,147 @@ export const Picker: React.FC<IPickerProps> = ({
       return itemName.includes(search);
     });
   }, [searchPhrase, options]);
+  const hasItems = items.length > 0;
 
-  const selectedItemsKeys = React.useMemo(() => {
-    if (!selected) {
-      return null;
+  const { refs, floatingStyles, context, isPositioned } =
+    useFloating<HTMLButtonElement>({
+      nodeId,
+      open,
+      strategy: floatingStrategy,
+      onOpenChange: (open) => {
+        setOpen(open);
+        setSearchPhrase(null);
+      },
+      whileElementsMounted: autoUpdate,
+      middleware: [
+        offset(8),
+        flip({ padding: 10 }),
+        shift(),
+        floatingSize({
+          apply({ availableHeight, rects, elements }) {
+            ReactDOM.flushSync(() => {
+              setMaxHeight(availableHeight);
+            });
+            Object.assign(elements.floating.style, {
+              width: `${rects.reference.width}px`,
+            });
+          },
+          padding: overflowPadding,
+        }),
+      ],
+    });
+
+  const click = useClick(context, { enabled: !disabled, ...useClickHookProps });
+  const role = useRole(context, { role: 'listbox' });
+  const dismiss = useDismiss(context, useDismissHookProps);
+  const listNavigation = useListNavigation(context, {
+    enabled: hasItems && !disabled,
+    listRef: listElementsRef,
+    activeIndex,
+    onNavigate: setActiveIndex,
+    virtual: true,
+    disabledIndices: findIndicesWhere(
+      items,
+      (item) => !!item.disabled || !!item.groupHeader
+    ),
+  });
+
+  const { getReferenceProps, getFloatingProps, getItemProps } = useInteractions(
+    [click, dismiss, role, listNavigation]
+  );
+
+  const handleSelect = (key: string) => {
+    if (type === 'single') {
+      setOpen(false);
+      setSelectedKeys(() => {
+        onSelect(options.filter((item) => item.key === key));
+
+        return [key];
+      });
+    } else {
+      setSelectedKeys((prev) => {
+        const newSelectedIndices = prev.includes(key)
+          ? prev.filter((i) => i !== key)
+          : [...prev, key];
+
+        const newSelectedItems = options.filter((item) =>
+          newSelectedIndices.includes(item.key)
+        );
+
+        onSelect(newSelectedItems);
+
+        return newSelectedIndices;
+      });
     }
+  };
 
-    return selected.map((item) => item.key);
-  }, [selected]);
+  const handleOnFilter = (text: string) => setSearchPhrase(text);
+
+  const handleItemRemove = (itemKey: string) => {
+    handleSelect(itemKey);
+  };
+
+  const handleClear = () => {
+    setOpen(false);
+    setSelectedKeys([]);
+    onSelect(null);
+  };
 
   return (
-    <div ref={triggerRef} className={mergedClassNames} id={props.id}>
-      <div className={styles[`${baseClass}__container`]}>
-        <Trigger
-          testId={props['data-testid']}
+    <div id={id} className={cx(styles['picker-wrapper'], className)}>
+      <PickerTrigger
+        getReferenceProps={getReferenceProps}
+        setReference={refs.setReference}
+        testId={props['data-testid']}
+        isItemSelected={selectedKeys.length > 0}
+        isOpen={open}
+        onClear={handleClear}
+        hideClearButton={hideClearButton}
+        isDisabled={disabled}
+        isError={error}
+        isRequired={isRequired}
+        isMultiSelect={type === 'multi'}
+        size={size}
+      >
+        <PickerTriggerBody
+          isOpen={open}
           isSearchDisabled={searchDisabled}
-          isError={error}
-          isOpen={isListOpen}
           isDisabled={disabled}
-          isItemSelected={!!selected}
-          isRequired={isRequired}
-          isMultiSelect={type === 'multi'}
+          placeholder={placeholder}
+          selectedItems={selected}
+          type={type}
           size={size}
-          hideClearButton={hideClearButton}
-          onTrigger={handleTrigger}
-          onClear={handleClear}
-        >
-          <TriggerBody
-            isOpen={isListOpen}
-            isSearchDisabled={searchDisabled}
-            isDisabled={disabled}
-            placeholder={placeholder}
-            items={selected}
-            type={type}
-            size={size}
-            clearSearchAfterSelection={clearSearchAfterSelection}
-            onItemRemove={handleItemRemove}
-            onFilter={handleOnFilter}
-          />
-        </Trigger>
-        <PickerList
-          selectedItemsKeys={selectedItemsKeys}
-          items={items}
-          isOpen={isListOpen}
-          isMultiSelect={type === 'multi'}
-          emptyStateText={noSearchResultText}
-          selectAllOptionText={selectAllOptionText}
-          onClose={handleOnClose}
-          onSelect={handleSelect}
-          onSelectAll={handleSelectAll}
+          clearSearchAfterSelection={clearSearchAfterSelection}
+          onItemRemove={handleItemRemove}
+          onFilter={handleOnFilter}
         />
-      </div>
+      </PickerTrigger>
+      <FloatingNode id={nodeId}>
+        {open && (
+          <FloatingPortal>
+            <PickerList
+              pickerType={type}
+              options={items}
+              listClassName={listClassName}
+              context={context}
+              setFloating={refs.setFloating}
+              floatingStyles={floatingStyles}
+              maxHeight={maxHeight}
+              isPositioned={isPositioned}
+              pointer={pointer}
+              activeIndex={activeIndex}
+              selectedKeys={selectedKeys}
+              listElementsRef={listElementsRef}
+              setPointer={setPointer}
+              onSelect={handleSelect}
+              getFloatingProps={getFloatingProps}
+              getItemProps={getItemProps}
+              emptyStateText={noSearchResultText}
+              selectAllOptionText={selectAllOptionText}
+            />
+          </FloatingPortal>
+        )}
+      </FloatingNode>
     </div>
   );
 };
