--- packages/react-components/src/components/Picker/PickerList.module.scss
@@ -1,166 +1,0 @@
-@import '../../utils/StackingContextLevel';
-
-$base-class: 'picker-list';
-
-.list-wrapper {
-  position: absolute;
-  top: calc(100% + 4px);
-  z-index: $stacking-context-level-dropdown;
-  border-radius: var(--radius-3);
-  box-shadow: var(--shadow-pop-over);
-  width: 100%;
-  overflow: auto;
-}
-
-.#{$base-class} {
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-05);
-  margin: 0;
-  background-color: var(--picker-list-background);
-  padding: var(--spacing-2);
-  max-height: 242px;
-  overflow: auto;
-  line-height: 22px;
-  font-size: 15px;
-  font-weight: 400;
-  color-scheme: var(--color-scheme);
-
-  &__no-results {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 36px;
-  }
-
-  &__item {
-    box-sizing: border-box;
-    display: flex;
-    position: relative;
-    align-items: center;
-    justify-content: space-between;
-    border: 1px solid transparent;
-    border-radius: var(--radius-3);
-    cursor: pointer;
-    padding: 6px 11px;
-    height: 100%;
-
-    &:hover {
-      background-color: var(--picker-list-option-background-hover);
-    }
-
-    &:focus,
-    &:active {
-      background-color: var(--picker-list-option-background-focus);
-    }
-
-    &[aria-current='true'] {
-      border-color: var(--action-primary-default);
-      background-color: var(--picker-list-option-background-focus);
-    }
-
-    &[aria-disabled='true'] {
-      color: var(--content-basic-disabled);
-
-      &:hover {
-        background-color: unset;
-        cursor: auto;
-      }
-    }
-
-    &[aria-selected='true'] {
-      background-color: var(--picker-list-option-background-active);
-
-      &:hover {
-        background-color: var(--picker-list-option-background-active-hover);
-      }
-    }
-
-    &--select-all {
-      border-bottom: 1px solid var(--border-basic-tertiary);
-    }
-
-    &__header {
-      display: flex;
-      position: relative;
-      align-items: center;
-      justify-content: space-between;
-      z-index: 1;
-      background-color: var(--picker-list-group-background);
-      cursor: auto;
-      padding: 12px 12px 4px;
-      height: 36px;
-      text-transform: uppercase;
-      color: var(--content-basic-secondary);
-      font-size: 12px;
-      font-weight: 600;
-
-      &::after {
-        position: absolute;
-        top: 0;
-        right: -8px;
-        left: -8px;
-        z-index: -1;
-        background: var(--picker-list-group-background);
-        height: 36px;
-        content: '';
-      }
-
-      &:nth-of-type(1)::before {
-        position: absolute;
-        top: calc(var(--spacing-2) * -1);
-        right: -8px;
-        left: -8px;
-        z-index: -1;
-        background: var(--picker-list-group-background);
-        height: var(--spacing-2);
-        content: '';
-      }
-    }
-
-    &__content {
-      display: flex;
-      flex: 1;
-      align-items: center;
-      overflow: hidden;
-    }
-
-    &__custom {
-      width: 100%;
-      height: auto;
-    }
-
-    &__checkbox {
-      align-self: center;
-      margin-right: var(--spacing-2);
-    }
-
-    &__avatar,
-    &__icon {
-      align-self: center;
-      margin-right: var(--spacing-2);
-      border-radius: var(--radius-3);
-      width: 20px;
-      height: 20px;
-      color: var(--content-basic-primary);
-    }
-
-    &__label-container {
-      display: flex;
-      flex: 1;
-      flex-direction: column;
-      justify-content: center;
-      overflow: hidden;
-    }
-
-    &__main-label {
-      font-weight: bold;
-    }
-
-    &__secondary-label {
-      color: var(--content-basic-secondary);
-      font-size: 12px;
-    }
-  }
-}
