+++ packages/react-components/src/components/Picker/components/PickerTrigger.module.scss
