+++ packages/react-components/src/components/Picker/components/PickerTriggerBody.module.scss
