--- packages/react-components/src/components/PickerV2/Picker.module.scss
@@ -1,3 +1,0 @@
-.picker-wrapper {
-  width: 100%;
-}
