--- packages/react-components/src/components/PickerV2/components/PickerTriggerBody.module.scss
@@ -1,89 +1,0 @@
-$base-class: 'picker-trigger-body';
-
-.#{$base-class} {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 4px;
-  justify-content: flex-start;
-  max-width: 100%;
-
-  &--single {
-    flex-wrap: nowrap;
-  }
-
-  &__item-container {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 4px;
-    max-height: 96px;
-    overflow-y: auto;
-    color-scheme: var(--color-scheme);
-  }
-
-  &__item {
-    display: flex;
-    flex: 1;
-    align-items: center;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    &__content {
-      text-wrap: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-
-    &__avatar,
-    &__icon {
-      align-self: center;
-      margin-right: var(--spacing-2);
-      border-radius: 50%;
-      width: 20px;
-      height: 20px;
-    }
-  }
-
-  &__input {
-    display: flex;
-    flex-grow: 2;
-    align-items: center;
-    outline: none;
-    border: 0;
-    background: none;
-    padding: 0;
-    height: 22px;
-    line-height: 22px;
-    color: var(--content-basic-primary);
-    font-size: 15px;
-
-    &--compact {
-      height: 20px;
-    }
-
-    &--medium {
-      height: 24px;
-    }
-
-    &--large {
-      height: 32px;
-    }
-  }
-
-  &__tag {
-    border-radius: var(--radius-1);
-    max-width: 100%;
-
-    &--compact {
-      height: 20px;
-    }
-
-    &--medium {
-      height: 24px;
-    }
-
-    &--large {
-      height: 32px;
-    }
-  }
-}
