--- packages/react-components/src/components/PickerV2/index.ts
@@ -1,3 +1,0 @@
-export * from './Picker';
-export * from './components/PickerList';
-export type { PickerType, IPickerListItem, IPickerProps } from './types';
